--- a/go/mysql/collations/uca.go
+++ b/go/mysql/collations/uca.go
@@ -100,7 +100,7 @@ nextLevel:
 			if l != r || !lok || !rok {
 				break
 			}
-			if fastleft.Level() != level || fastleft.Level() != level {
+			if fastleft.Level() != level || fastright.Level() != level {
 				break
 			}
 		}