--- a/go/vt/vtgate/evalengine/comparisons.go
+++ b/go/vt/vtgate/evalengine/comparisons.go
@@ -17,6 +17,10 @@ limitations under the License.
 package evalengine
 
 import (
+	"bytes"
+	"fmt"
+	"sync"
+
 	"vitess.io/vitess/go/mysql/collations"
 	"vitess.io/vitess/go/sqltypes"
 	vtrpcpb "vitess.io/vitess/go/vt/proto/vtrpc"
@@ -382,4 +386,243 @@ func (l *LikeExpr) typeof(env *ExpressionEnv) (sqltypes.Type, flag) {
 	_, f1 := l.Left.typeof(env)
 	_, f2 := l.Right.typeof(env)
 	return sqltypes.Int64, f1 | f2
+}
+
+// UnsupportedComparisonError represents the error where the comparison between the two types is unsupported on vitess
+type UnsupportedComparisonError struct {
+	Type1 sqltypes.Type
+	Type2 sqltypes.Type
+}
+
+// Error function implements the error interface
+func (err UnsupportedComparisonError) Error() string {
+	return fmt.Sprintf("types are not comparable: %v vs %v", err.Type1, err.Type2)
+}
+
+// UnsupportedCollationError represents the error where the comparison using provided collation is unsupported on vitess
+type UnsupportedCollationError struct {
+	ID collations.ID
+}
+
+// Error function implements the error interface
+func (err UnsupportedCollationError) Error() string {
+	return fmt.Sprintf("cannot compare strings, collation is unknown or unsupported (collation ID: %d)", err.ID)
+}
+
+// NullsafeCompare returns 0 if v1==v2, -1 if v1<v2, and 1 if v1>v2.
+// NULL is the lowest value. If any value is
+// numeric, then a numeric comparison is performed after
+// necessary conversions. If none are numeric, then it's
+// a simple binary comparison. Uncomparable values return an error.
+func NullsafeCompare(v1, v2 sqltypes.Value, collationID collations.ID) (int, error) {
+	// Based on the categorization defined for the types,
+	// we're going to allow comparison of the following:
+	// Null, isNumber, IsBinary. This will exclude IsQuoted
+	// types that are not Binary, and Expression.
+	if v1.IsNull() {
+		if v2.IsNull() {
+			return 0, nil
+		}
+		return -1, nil
+	}
+	if v2.IsNull() {
+		return 1, nil
+	}
+
+	if isByteComparable(v1.Type()) && isByteComparable(v2.Type()) {
+		v1Bytes, err1 := v1.ToBytes()
+		if err1 != nil {
+			return 0, err1
+		}
+		v2Bytes, err2 := v2.ToBytes()
+		if err2 != nil {
+			return 0, err2
+		}
+		return bytes.Compare(v1Bytes, v2Bytes), nil
+	}
+
+	typ, err := CoerceTo(v1.Type(), v2.Type()) // TODO systay we should add a method where this decision is done at plantime
+	if err != nil {
+		return 0, err
+	}
+
+	switch {
+	case typ == sqltypes.VarChar:
+		v1Bytes := v1.Raw()
+		v2Bytes := v2.Raw()
+
+		if collationID == collations.Unknown {
+			return 0, UnsupportedCollationError{ID: collationID}
+		}
+
+		collation := collations.Local().LookupByID(collationID)
+		if collation == nil {
+			return 0, UnsupportedCollationError{ID: collationID}
+		}
+
+		switch result := collation.Collate(v1Bytes, v2Bytes, false); {
+		case result < 0:
+			return -1, nil
+		case result > 0:
+			return 1, nil
+		default:
+			return 0, nil
+		}
+
+	case sqltypes.IsNumber(typ):
+		v1cast := borrow()
+		v2cast := borrow()
+
+		defer func() {
+			erpool.Put(v1cast)
+			erpool.Put(v2cast)
+		}()
+
+		if err := v1cast.setValueCast(v1, typ); err != nil {
+			return 0, err
+		}
+		if err := v2cast.setValueCast(v2, typ); err != nil {
+			return 0, err
+		}
+		return compareNumeric(v1cast, v2cast)
+
+	default:
+		return 0, UnsupportedComparisonError{Type1: v1.Type(), Type2: v2.Type()}
+	}
+}
+
+// isByteComparable returns true if the type is binary or date/time.
+func isByteComparable(typ sqltypes.Type) bool {
+	if sqltypes.IsBinary(typ) {
+		return true
+	}
+	switch typ {
+	case sqltypes.Timestamp, sqltypes.Date, sqltypes.Time, sqltypes.Datetime, sqltypes.Enum, sqltypes.Set, sqltypes.TypeJSON, sqltypes.Bit:
+		return true
+	}
+	return false
+}
+
+// Min returns the minimum of v1 and v2. If one of the
+// values is NULL, it returns the other value. If both
+// are NULL, it returns NULL.
+func Min(v1, v2 sqltypes.Value, collation collations.ID) (sqltypes.Value, error) {
+	return minmax(v1, v2, true, collation)
+}
+
+// Max returns the maximum of v1 and v2. If one of the
+// values is NULL, it returns the other value. If both
+// are NULL, it returns NULL.
+func Max(v1, v2 sqltypes.Value, collation collations.ID) (sqltypes.Value, error) {
+	return minmax(v1, v2, false, collation)
+}
+
+func minmax(v1, v2 sqltypes.Value, min bool, collation collations.ID) (sqltypes.Value, error) {
+	if v1.IsNull() {
+		return v2, nil
+	}
+	if v2.IsNull() {
+		return v1, nil
+	}
+
+	n, err := NullsafeCompare(v1, v2, collation)
+	if err != nil {
+		return sqltypes.NULL, err
+	}
+
+	// XNOR construct. See tests.
+	v1isSmaller := n < 0
+	if min == v1isSmaller {
+		return v1, nil
+	}
+	return v2, nil
+}
+
+var erpool = sync.Pool{
+	New: func() interface{} {
+		return &EvalResult{}
+	},
+}
+
+func borrow() *EvalResult {
+	er := erpool.Get().(*EvalResult)
+	er.flags_ = 0
+	return er
+}
+
+func NullsafeCompare_old(v1, v2 sqltypes.Value, collationID collations.ID) (int, error) {
+	// Based on the categorization defined for the types,
+	// we're going to allow comparison of the following:
+	// Null, isNumber, IsBinary. This will exclude IsQuoted
+	// types that are not Binary, and Expression.
+	if v1.IsNull() {
+		if v2.IsNull() {
+			return 0, nil
+		}
+		return -1, nil
+	}
+	if v2.IsNull() {
+		return 1, nil
+	}
+
+	if isByteComparable(v1.Type()) && isByteComparable(v2.Type()) {
+		v1Bytes, err1 := v1.ToBytes()
+		if err1 != nil {
+			return 0, err1
+		}
+		v2Bytes, err2 := v2.ToBytes()
+		if err2 != nil {
+			return 0, err2
+		}
+		return bytes.Compare(v1Bytes, v2Bytes), nil
+	}
+
+	typ, err := CoerceTo(v1.Type(), v2.Type()) // TODO systay we should add a method where this decision is done at plantime
+	if err != nil {
+		return 0, err
+	}
+	var v1cast, v2cast EvalResult
+	if err := v1cast.setValueCast(v1, typ); err != nil {
+		return 0, err
+	}
+	if err := v2cast.setValueCast(v2, typ); err != nil {
+		return 0, err
+	}
+
+	if sqltypes.IsNumber(typ) {
+		return compareNumeric(&v1cast, &v2cast)
+	}
+	if sqltypes.IsText(typ) || sqltypes.IsBinary(typ) {
+		if collationID == collations.Unknown {
+			return 0, UnsupportedCollationError{
+				ID: collationID,
+			}
+		}
+		collation := collations.Local().LookupByID(collationID)
+		if collation == nil {
+			return 0, UnsupportedCollationError{
+				ID: collationID,
+			}
+		}
+		v1Bytes, err1 := v1.ToBytes()
+		if err1 != nil {
+			return 0, err1
+		}
+		v2Bytes, err2 := v2.ToBytes()
+		if err2 != nil {
+			return 0, err2
+		}
+		switch result := collation.Collate(v1Bytes, v2Bytes, false); {
+		case result < 0:
+			return -1, nil
+		case result > 0:
+			return 1, nil
+		default:
+			return 0, nil
+		}
+	}
+	return 0, UnsupportedComparisonError{
+		Type1: v1.Type(),
+		Type2: v2.Type(),
+	}
 }